requirements.txt
Django==4.2.7
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
